chore: unifies raw and template manifest types
The raw and template manifest structs differ only in their content processing methods.

For templates, Golang HTML templating is applied before converting the content to Unstructured objects.

Although we could theoretically treat all manifests as templates, we will adhere to the convention of using a '.tmpl.' in the file name to indicate a template and only then process it as such.

This PR consolidates the two implementations into one, reducing duplication and resulting in cleaner code.