feat: reworks Kustomize manifests to support transfomers
This way we can decouple kustomize core logic from namespace and labels plugins
which now can be supplied from the caller instead.

Feature can have multiple KustomizeManifests defined through subsequent
calls of Kustomize() sub-builder. There is a possibility to define
global list of plugins, which will be applied for each of those. This is
used by the FeatureHandler to ensure both target namespace and component labels (if
applicable) will be added to kustomize resources before applied to the
cluster.

This change also restructures the Feature builder and next to
Kustomize() sub-builder it also introduces Manifests() for regular YAML
files or templates without Kustomize engine involved.